Fix Kubernetes client helper lint issues
